Move cliqz antitracking and adblocking data retrieval to utils functions. Converted BrowserButton to use them
