+++ src/classes/BrowserButton.js
@@ -18,6 +18,7 @@ import foundBugs from './FoundBugs';
 import cliqz from './Cliqz';
 import rewards from './Rewards';
 import Policy from './Policy';
+import { getCliqzAntitrackingData, getCliqzAdblockingData } from '../utils/cliqzModuleData';
 import { getTab } from '../utils/utils';
 import { log } from '../utils/common';
 import globals from './Globals';
@@ -146,12 +147,12 @@ class BrowserButton {
 			return;
 		}
 
-		this._getAntiTrackCount(tabId).then((antiTrackingCount) => {
+		getCliqzAntitrackingData(tabId).then((antitrackingData) => {
 			const { appsCount, appsAlertCount } = this._getTrackerCount(tabId);
-			const adBlockingCount = this._getAdBlockCount(tabId);
+			const adBlockingCount = getCliqzAdblockingData(tabId).totalCount;
 
 			alert = (appsAlertCount > 0);
-			trackerCount = (appsCount + antiTrackingCount + adBlockingCount).toString();
+			trackerCount = (appsCount + antitrackingData.totalUnsafeCount + adBlockingCount).toString();
 
 			// gray-out the icon when blocking has been disabled for whatever reason
 			if (trackerCount === '') {
@@ -174,52 +175,6 @@ class BrowserButton {
 			appsCount: apps.all,
 			appsAlertCount: apps.total,
 		};
-	}
-
-	/**
-	 * Get tracker count for Anti Tracking in a promise
-	 * @param  {number} tabId  the Tab Id
-	 * @return {Promise}       the number of trackers as a Promise
-	 */
-
-// TODO restructure this to eliminiate duplication of effort and data structures
-// between this class, PanelData, and the summary reducer on the front end
-// probably, this means building out a CliqzModuleData class real quick
-	_getAntiTrackCount(tabId) {
-		return new Promise((resolve) => {
-			if (!conf.enable_anti_tracking || !antitracking.background) {
-				resolve(0);
-			}
-			antitracking.background.actions.aggregatedBlockingStats(tabId).then((antiTracking) => {
-				let totalUnsafeCount = 0;
-				for (const category in antiTracking) {
-					if (antiTracking.hasOwnProperty(category)) {
-						for (const app in antiTracking[category]) {
-							if (antiTracking[category][app] === 'blocked') { // until 2/17/19 we were checking for 'unsafe'
-								totalUnsafeCount++;
-							}
-						}
-					}
-				}
-				resolve(totalUnsafeCount);
-			}).catch(() => {
-				// if we encounter an error, return 0
-				resolve(0);
-			});
-		});
-	}
-
-	/**
-	 * Get tracker count for Ad Blocking
-	 * @param  {number} tabId  the Tab Id
-	 * @return {number}        the number of trackers in an object
-	 */
-	_getAdBlockCount(tabId) {
-		if (!conf.enable_ad_block || !adblocker.background) {
-			return 0;
-		}
-		const adBlocking = adblocker.background.actions.getAdBlockInfoForTab(tabId);
-		return adBlocking && adBlocking.totalCount || 0;
 	}
 }
